evidence_preservation/requirements.txt
boto3
azure-monitor-query
azure-identity
google-cloud-logging
cryptography
requests
python-dotenv
asn1crypto
